refactor excel pkg; polish related codes
