--- pkg/excel/decode_test.go
@@ -1,37 +1,0 @@
-// Copyright (c) 2021 Terminus, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//      http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package excel
-
-//import (
-//	"os"
-//	"testing"
-//
-//	"github.com/davecgh/go-spew/spew"
-//	"github.com/stretchr/testify/assert"
-//)
-//
-//func TestDecode(t *testing.T) {
-//	f, err := os.Open("/tmp/测试用例.xlsx")
-//	assert.NoError(t, err)
-//	data, err := Decode(f)
-//	assert.NoError(t, err)
-//	spew.Dump(data)
-//
-//	f, err = os.Open("/tmp/测试用例2.xlsx")
-//	assert.NoError(t, err)
-//	data, err = Decode(f)
-//	assert.NoError(t, err)
-//	spew.Dump(data)
-//}
